+++ src/collective/collectionfilter/filteritems.py
@@ -179,7 +179,9 @@ def get_filter_items(
         k: v for k, v in urlquery.items() if k not in (idx, idx + '_op')
     }
     ret = [{
-        'title': translate(_('subject_all', default=u'All'), context=getRequest()),
+        'title': translate(
+            _('subject_all', default=u'All'), context=getRequest()
+        ),
         'url': u'{0}/?{1}'.format(
             collection_url,
             urlencode(safe_encode(urlquery_all), doseq=True)