+++ src/collective/collectionfilter/portlets/collectionfilter.py
@@ -5,6 +5,7 @@ from ..interfaces import ICollectionFilterSchema
 from ..vocabularies import DEFAULT_FILTER_TYPE
 from plone.app.portlets.portlets import base
 from plone.portlets.interfaces import IPortletDataProvider
+from Products.CMFPlone.utils import get_top_request
 from Products.CMFPlone.utils import getFSVersionTuple
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope.interface import implementer
@@ -79,7 +80,8 @@ class Renderer(BaseFilterView, base.Renderer):
 
     @property
     def id(self):
-        portlethash = self.request.form.get(
+        request = get_top_request(self.request)
+        portlethash = request.form.get(
             'portlethash',
             getattr(self, '__portlet_metadata__', {}).get('hash', '')
         )